requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
bleach==6.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-apscheduler==0.6.2
django-cleanup==8.1.0
django-filter==24.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-summernote==0.8.20.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
Markdown==3.6
packaging==24.1
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
